+++ src/main.js
@@ -1,10 +1,10 @@
 import { createApp } from 'vue'
 import { createPinia } from 'pinia'
-import piniaPluginPersistedstate from "pinia-plugin-persistedstate"
-
 import App from './App.vue'
+import './registerServiceWorker'
 import router from './router'
-//import { useRoute } from 'vue-router'
+import piniaPluginPersistedstate from 'pinia-plugin-persistedstate'
+import './index.css'
 
 const app = createApp(App)
 const pinia = createPinia()