+++ src/router/index.js
@@ -1,40 +1,40 @@
 import { createRouter, createWebHistory } from 'vue-router'
-import Main from '../views/Main.vue'
-import Story from '../views/Story.vue'
+import MainView from '../views/MainView.vue'
+import StoryView from '../views/StoryView.vue'
 import StoryEdit from '../views/StoryEdit.vue'
 import GlobalSettingStory from '../views/GlobalSettingStory.vue'
 import PlayStory from '../views/PlayStory.vue'
 
 const router = createRouter({
-  history: createWebHistory(import.meta.env.BASE_URL),
+  history: createWebHistory(process.env.BASE_URL),
   routes: [
     {
       path: '/',
       name: 'home',
-      component: Main
+      component: MainView,
     },
     {
       path: '/story/:id',
       name: 'story',
-      component: Story
+      component: StoryView,
     },
     {
       path: '/story/:id/edit/:pid',
       name: 'edit story',
-      component: StoryEdit
+      component: StoryEdit,
     },
     {
       path: '/story/:id/global',
       name: 'global setting',
-      component: GlobalSettingStory
+      component: GlobalSettingStory,
     },
     {
       path: '/story/:id/play',
       name: 'play story',
-      component: PlayStory
+      component: PlayStory,
     },
-    { path: '/:pathMatch(.*)*', name: 'NotFound', component: Main }
-  ]
+    { path: '/:pathMatch(.*)*', name: 'NotFound', component: MainView },
+  ],
 })
 
 export default router