+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 ## 0.0.1
+
 Date : 17-April-2022
 
 > The Very First Dev Of This Tools.
 
 List Addition / Changes :
+
 - Edit Existing Story.
   - Edit Title Story.
   - Generate IFID.
@@ -16,11 +18,13 @@ List Addition / Changes :
 - Run Edited Story.
 
 ## 0.1.0
+
 Date : 23-April-2022
 
 > Many Changes. Also Dark Theme !.
 
 List Addition / Changes :
+
 - Add, Edit, And Delete Story.
 - Backup And Restore Some Story Or All Of Them.
 - Remodeling The Layout.
@@ -29,51 +33,79 @@ List Addition / Changes :
 - Dark Theme. Will Not Use Light. Light Attract Bugs. :3
 
 ## 0.1.1
+
 Date : 23-April-2022
 
 > Small Changes.
 
 List Addition / Changes :
+
 - Now You Can Reorder The List Of Story And Passages.
 - Fixing Some Styling.
 
 ## 0.1.2
+
 Date : 24-April-2022
 
 > Small Changes.
 
 List Addition / Changes :
+
 - Removed "Generate IFID" Button. Since IFID Is Unique ID It Should Only Generate When Creating Story.
 - Fixing Some Styling.
 - Fixing Some 404 Error Not Found. Hope This Works.
 
 ## 0.1.2.X
+
 Date : 25-April-2022 Until 30-April-2022
 
 > Trying New Feature. Wait Till Release.  
 > I Think I Must Commit Daily. :3.
 
 List Addition / Changes :
+
 - Tying New Styling.
 - Test For New Story Formats.
 - Test For "How Big The Game Script This Tools Can Handle ?". I Think It's Based On Your Browser. But I Think It Can Handle A LOT.
 
 ## 1.0.0
+
 Date : 30-April-2022
 
 > Big Changes. Finally, I Can Use This Tools To It's Full Potentials. :3.
 
 List Addition / Changes :
+
 - Few New Styling.
 - Now Support All Official Listed Story Formats On Twine 2. Such As Chapbook, Harlowe, Snowman, And Sugarcube.
 - Now Wrap Text Setting Already Made Globally.
 
 ## 1.0.1
+
 Date : 16-May-2022
 
 > Small Changes.
 
 List Addition / Changes :
+
 - Few Fix Styling
 - Edit Passage Names And Story Title Now No Need To Press The "Set" Button.
 - Preparing For Exporting To .apk
+
+## 1.1.0
+
+Date : 29-October-2022
+
+> Very Big Changes !.  
+> Sorry For Abandoning This Project. :3.  
+> Now I'm Back With Very Big Updates !.
+
+List Addition / Changes :
+
+- Redesign The App Layout. Now It Looks Clean And Modern Thanks To Daisy UI. :)
+- App Theme Will Follow Your Device Theme.
+- Play And Export Story Now With Just 1 Button. Easy !. :3.
+- Add Ability To Search The Story Names / Passage Names.
+- Edit Passage Names Now Need To Press The "Set" Button.
+- Upgrade The Version Of Vue.
+- Now This App Is A PWA !. U Can Install This App To Ur Device If You Open This App With Chrome Browser !. So U Can Use This App Without Internet !.